Improve the backup image

+++ backup/entrypoint.sh
@@ -8,39 +8,41 @@ IFS=$'\n\t'
 AWS_S3_BUCKET="${AWS_S3_BUCKET:-backup_$(date +%s | sha256sum | base64 | head -c 16 ; echo)}"
 AWS_DEFAULT_REGION="${AWS_DEFAULT_REGION:-us-east-1}"
 GPG_RECIPIENT="${GPG_RECIPIENT:-}"
-BACKUP_PATHS="${BACKUP_PATHS:-/backup}"
-BACKUP_RESTORE="${BACKUP_RESTORE:-false}"
+GPG_KEYS_PATH="${GPG_KEYS_PATH:-/keys}"
+BACKUP_PATH="${BACKUP_PATH:-/backup}"
+CRON_TIME="${CRON_TIME:-8 */8 * * *}"
 NOW=$(date +"%Y-%m-%d_%H-%M-%S")
 TMPDIR=$(mktemp -d 2>/dev/null || mktemp -d -t 'tmp')
-BACKUP_FILE="${NOW}.tar.xz"
-BACKUP_FILE_ENCRYPTED="${BACKUP_FILE}.gpg"
-RESTORE_FILE=
-RESTORE_FILE_DECRYPTED="decrypted.tar.xz"
-CRON_TIME="${CRON_TIME:-8 */8 * * *}"
 
 # Log message
 log(){
   echo "$(date): ${*}"
 }
 
-# Tar GZ everything
-create_archive(){
-  log "Create ${BACKUP_FILE}"
-  tar cJf "$BACKUP_FILE" "$BACKUP_PATHS"
-}
-
 # Import public GPG key
 import_gpg_keys(){
-  if [[ -d /keys ]]; then
+  if [[ "$(ls -A "$GPG_KEYS_PATH" 2>&1)" ]]; then
     log "Import all keys in '/keys' folder"
-    gpg --batch --import /keys/*
+    gpg --batch --import ${GPG_KEYS_PATH}/*
+  fi
+}
+
+# Tar GZ the backup path
+create_archive(){
+  export BACKUP_FILE="${NOW}.tar.xz"
+  if [[ "$(ls -A "$BACKUP_PATH" 2>&1)" ]]; then
+    log "Create ${BACKUP_FILE}"
+    tar cJf "$BACKUP_FILE" -C "$BACKUP_PATH" .
   fi
 }
 
 # Encrypt the backup tar.gz
 encrypt_archive(){
-  log "Encrypt ${BACKUP_FILE_ENCRYPTED}"
-  gpg --trust-model always --output "$BACKUP_FILE_ENCRYPTED" --batch --encrypt --recipient "$GPG_RECIPIENT" "$BACKUP_FILE"
+  export BACKUP_FILE_ENCRYPTED="${BACKUP_FILE}.gpg"
+  if [[ -n "$GPG_RECIPIENT" ]] && [[ -s "$BACKUP_FILE" ]]; then
+    log "Encrypt ${BACKUP_FILE_ENCRYPTED}"
+    gpg --trust-model always --output "$BACKUP_FILE_ENCRYPTED" --batch --encrypt --recipient "$GPG_RECIPIENT" "$BACKUP_FILE"
+  fi
 }
 
 # Create bucket, if it doesn't already exist
@@ -59,24 +61,27 @@ ensure_s3_bucket(){
 # Upload archive to AWS S3
 upload_archive(){
   # Copy to AWS S3
-  aws s3 cp "$BACKUP_FILE_ENCRYPTED" \
-    "s3://${AWS_S3_BUCKET}/${BACKUP_PREFIX}/${BACKUP_FILE_ENCRYPTED}"
+  if [[ -s "$BACKUP_FILE_ENCRYPTED" ]]; then
+    aws s3 cp "$BACKUP_FILE_ENCRYPTED" \
+      "s3://${AWS_S3_BUCKET}/${BACKUP_PREFIX}/${BACKUP_FILE_ENCRYPTED}"
+  elif [[ -s "$BACKUP_FILE" ]]; then
+    aws s3 cp "$BACKUP_FILE" \
+      "s3://${AWS_S3_BUCKET}/${BACKUP_PREFIX}/${BACKUP_FILE}"
+  fi
 }
 
-# Backup logic
+# Backup archive
 backup_archive(){
   export BACKUP_PREFIX=${1:-hourly}
 
-  ensure_s3_bucket
   create_archive
-  if [[ -n "$GPG_RECIPIENT" ]]; then
-    encrypt_archive
-  fi
+  encrypt_archive
+  ensure_s3_bucket
   upload_archive
 }
 
 # List the latest S3 archive
-latest_s3_archive(){
+get_latest_s3_archive(){
   if ! aws s3api list-objects \
     --bucket "$AWS_S3_BUCKET" \
     --query 'reverse(sort_by(Contents, &LastModified))[0].Key' \
@@ -88,22 +93,35 @@ latest_s3_archive(){
 
 # Download the latest archive from S3
 download_s3_archive(){
-  RESTORE_FILE=$(latest_s3_archive)
   aws s3 cp "s3://${AWS_S3_BUCKET}/${RESTORE_FILE}" "$RESTORE_FILE"
 }
 
 # Decrypt archive
 decrypt_archive(){
-  log "Decrypt ${RESTORE_FILE}"
-  export GPG_TTY=/dev/console
-  gpg --batch --output "$RESTORE_FILE_DECRYPTED" --decrypt "$RESTORE_FILE"
+  export RESTORE_FILE_DECRYPTED="decrypted.tar.xz"
+  if [[ -s "$RESTORE_FILE" ]]; then
+    log "Decrypt ${RESTORE_FILE}"
+    export GPG_TTY=/dev/console
+    gpg --batch --output "$RESTORE_FILE_DECRYPTED" --decrypt "$RESTORE_FILE"
+  fi
 }
 
 # Extract the latest archive
 extract_archive(){
-  log "Extract '${RESTORE_FILE_DECRYPTED}' to '/restore'"
-  mkdir -p /restore
-  tar xJf "$RESTORE_FILE_DECRYPTED" --directory /restore
+  if [[ -s "$RESTORE_FILE_DECRYPTED" ]]; then
+    log "Extract '${RESTORE_FILE_DECRYPTED}' to '/restore'"
+    mkdir -p /restore
+    tar xJf "$RESTORE_FILE_DECRYPTED" --directory /restore
+  fi
+}
+
+# Restore archive
+restore_archive(){
+  export RESTORE_FILE="$(get_latest_s3_archive)"
+
+  download_s3_archive
+  decrypt_archive
+  extract_archive
 }
 
 # Remove archives
@@ -122,6 +140,9 @@ main(){
     # Trap exit
   trap 'bye $?' HUP INT QUIT TERM
 
+  # Import GPG keys
+  import_gpg_keys
+
   # Set working directory
   cd "$TMPDIR" || exit
 
@@ -129,22 +150,18 @@ main(){
   cmd="${1:-once}"
   case "$cmd" in
     once)
-      import_gpg_keys
       ensure_s3_bucket
       backup_archive
       ;;
     cron)
-      import_gpg_keys
-      ensure_s3_bucket
-
-      log 'Initial backup'
+      log 'Start initial backup'
       backup_archive 'hourly'
 
       cat > /etc/crontabs/root <<CRON
-${CRON_TIME} /entrypoint.sh hourly >> /backup.log 2>&1
-2 2 * * * /entrypoint.sh daily >> /backup.log 2>&1
-3 3 * * 6 /entrypoint.sh weekly >> /backup.log 2>&1
-5 5 1 * * /entrypoint.sh monthly >> /backup.log 2>&1
+${CRON_TIME} /entrypoint.sh hourly
+2 2 * * * /entrypoint.sh daily
+3 3 * * 6 /entrypoint.sh weekly
+5 5 1 * * /entrypoint.sh monthly
 CRON
       log "Run backups as a cronjob for ${CRON_TIME}"
       exec crond -l 2 -f
@@ -162,10 +179,8 @@ CRON
       backup_archive 'monthly'
       ;;
     restore)
-      import_gpg_keys
-      download_s3_archive
-      decrypt_archive
-      extract_archive
+      log 'Restoring files'
+      restore_archive
       ;;
     *)
       log "Unknown command: ${cmd}"; exit 1
